got rid of lots of warnings about not being able to find include files by removing includes in this format //include_once("database/dbMasterSchedule.php"); and replacing them with includes in tihs format include_once(dirname(_FILE_).'/../database/dbMasterSchedule.php
In addition, I got rid of the fatal error due to function name conflicts by renaming all conflicting functions in dbProjects.php. This is probably not the best fix - some of those functions should be refactored to a separate file and shared  - but we need to get the code running.